Consolidate file_reader.h, file_reader-internal.h
