+++ src/parquet/metadata-test.cc
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "parquet/file/metadata.h"
+#include "parquet/metadata.h"
 #include <gtest/gtest.h>
 #include "parquet/schema.h"
 #include "parquet/statistics.h"
